+++ Listener/AuthenticationListener.php
@@ -1,6 +1,6 @@
 <?php
 /*
- * (c) 2018: 975l <contact@975l.com>
+ * (c) 2018: 975L <contact@975l.com>
  * (c) 2018: Laurent Marquet <laurent.marquet@laposte.net>
  *
  * This source file is subject to the MIT license that is bundled
@@ -25,7 +25,7 @@ class AuthenticationListener implements EventSubscriberInterface
     private $container;
 
     public function __construct(
-        \Doctrine\ORM\EntityManagerInterface $em,
+        \Doctrine\ORM\EntityManager $em,
         \Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface $tokenStorage,
         \Symfony\Component\Security\Http\Authentication\AuthenticationUtils $authenticationUtils,
         \Symfony\Component\HttpFoundation\RequestStack $requestStack,